initial commit: basic check moves
Simple set up for ChessKit to know the rules of the board. The representation is a build-pattern based one, as it will be common to preserve the original position to evaluate different moves.

Underlying the current choice is to use a bit-board set up where each piece type has an associated bit-board per player. This bit-board can tell if there is a piece of certain type at some square on the board.

Each move results in either a `std::nullopt` which means that the move was illegal, or a new position object. On construction, a Position object is the starting position for chess with white to play.

With Position and Move, I expect the rule of chess can be covered entirely.