initial commit: The first move is White's
The Position object represents the state of the game and when default
constructed, it represents the starting position of a chess game.

This is the initial commit and as such it brings about the generic
project set up:

* uses CMake because it is the de-facto standard of C++ build systems.
* uses Catch2 because I like the expressiveness of the framework.
* uses Conan to pull dependencies (Catch2 in this case).
* it is set up to work C++17 as I found that the sweet spot of modern
  C++

The set up is meant to be used for TDD-style development where the unit
test can be run as the target `check-unit`.